+++ packages/theme-script/examples/vite/src/index.tsx
@@ -1,7 +1,9 @@
+/// <reference types="@hiogawa/theme-script/types" />
+// above "types" reference exposes __themeGet and __themeSet
+
 import "virtual:uno.css";
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { ThemeSelect } from "./theme-select";
 
 main();
 
@@ -21,6 +23,28 @@ function Root() {
         </div>
         <ThemeSelect />
       </div>
+    </div>
+  );
+}
+
+function ThemeSelect() {
+  const rerender = React.useReducer((prev) => !prev, true)[1];
+
+  return (
+    <div className="flex gap-2">
+      {["system", "light", "dark"].map((theme) => (
+        <button
+          key={theme}
+          className="flex-1 antd-btn antd-btn-default capitalize aria-selected:(text-colorPrimaryActive border-colorPrimaryActive)"
+          aria-selected={__themeGet() === theme}
+          onClick={() => {
+            __themeSet(theme);
+            rerender();
+          }}
+        >
+          {theme}
+        </button>
+      ))}
     </div>
   );
 }
